Adjust corert code for moving integer parsing/formatting to shared
- Deletes code that's moved to shared.
- Consolidates the rest without logic change into Number.CoreRT.cs.
- Fixes up a few functions to work with the changes in the shared code.
